+++ apps/opp-graphviz-app/main/Main.hs
@@ -1,4 +1,31 @@
 import Prelude
+-- import qualified Data.Yaml as Y
+import Persist.Mongo.Settings
+import Permissions.Onping.GraphVizWriter 
 
 main :: IO () 
-main = return () 
+main = do
+  eConf <- readDBConf "config.yml"
+  case eConf of 
+    (Left s) -> errorMessage s
+    (Right conf) -> do 
+               main' conf 
+
+main' :: MongoDBConf -> IO () 
+main' mdbc = do 
+  putStrLn "generatig dot file..."
+  writeGraphViz mdbc 
+  putStrLn "done generating dot file."
+  putStrLn "run with $> dot autograph.gv -T svg -o <outputname>.svg" 
+ 
+
+
+errorMessage :: String -> IO ()
+errorMessage s  = do
+	putStrLn "Error reading a config file."
+	putStrLn $ "Error: " ++ s
+
+
+
+
+