+++ core-libs/onping-permissions-core/changelog.md
@@ -11,18 +11,21 @@ Change Log
 
 *For complete history please view commit log*
 
-##Current Version 0.1.1
+##Current Version 0.1.3
 --------------------------	
 
 
-### Proposed Release 0.1.2
+### Proposed Release 0.1.4
 ----------------------------
 
 
 
 *Change History* 
 
-
++ 0.1.3
+  Changed runDB to runDB conf everywhere relevant (missed places last time)   
++ 0.1.2
+  Changed runDB to runDB conf everywhere relevant (missed places last time)   
 + 0.1.1
   Changed runDB to runDB conf everywhere relevant
 + 0.0.0